Fix EGIT_COMMIT support.
First of all, the variable has to be declared in optenv. Secondly,
git.eclass defaults it to the branch name, so we need to check it more
thoroughly.

Comparing it to EGIT_BRANCH may not be the best method to check if user
requested a particular commit (and not branch, for example) but that's
how the eclass does it and it should be sufficient for now.