Remove possible `getbinpkg` feature from the environment
My previous commit (e473be02e32d) tried fixing the issue of rebuilding binary
packages (they were fetched from the `$PORTAGE_BINHOST`), but forgot to also
remove the `getbinpkg` feature from the `FEATURES` (either set explicitly or in
`make.conf`).  This patch removes the feature if it is specified in the
environment and explicitly adds `-getbinpkg` to the feature list.

Signed-off-by: Martin Kletzander <nert.pinx@gmail.com>